Integration::new(): allow any type implementing HasRawDisplayHandle
Previously, to create an Integration an event loop must have been passed
as the first argument.
The only place where the event loop is used is in Integration::new(),
where it's uset to create an egui_winit::State, which accepts a
reference to any type implementing HasRawDisplayHandle.

By implementing this change, winit Windows can be used to create an
integration as well, which can be useful when an event loop isn't
available